+++ src/Enterspeed.Source.SitecoreCms.V9/Providers/EnterspeedSitecoreConfigurationProvider.cs
@@ -14,6 +14,6 @@ namespace Enterspeed.Source.SitecoreCms.V9.Providers
             _enterspeedConfigurationService = enterspeedConfigurationService;
         }
 
-        public EnterspeedConfiguration Configuration => _enterspeedConfigurationService.GetConfiguration();
+        public EnterspeedConfiguration Configuration => _enterspeedConfigurationService.GetConfigurationFromSitecore();
     }
 }