+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 * ParentId is now correctly mapped when parent represents the Site item.
 * Fixed ingest bug caused by redirects array not set to an instance.
 * Changed several field converters to have the correct properties added.
+* Moved configuration to Sitecore. Previous <setting /> tags are no longer used.
 
 ## [0.1.0 - 2021-03-22]